+++ src/library/scala/collection/immutable/HashSet.scala
@@ -96,7 +96,7 @@ class HashSet[A] extends Set[A]
     cached.iterator
   }
 
-  private def logLimit: Int = Math.sqrt(table.length).toInt
+  private def logLimit: Int = Math.sqrt(table.length.toDouble).toInt
 
   private def markUpdated(elem: A, del: Boolean) {
     val lv = loadFactor
